+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This file contains the repo-tags with the corresponding information.
 
+## [v1.1.0] - 2022-11-01
+ - add X-API-Key for auth @ eliona's APIv2
+ - auto websocket or websocket secure
 
 ## [v1.0.1] - 2022-10-24